# contracts/openapi/lifecycle-migration-campaigns-v1.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Lifecycle migration campaigns",
    "version": "1.0.0",
    "description": "Immutable schedules with current owner admission; observation publishers have separately mounted grants."
  },
  "paths": {
    "/v1/tenants/{tenant}/migration-campaigns": {
      "parameters": [
        {
          "name": "tenant",
          "in": "path",
          "required": true,
          "schema": {
            "type": "string",
            "format": "uuid"
          }
        }
      ],
      "get": {
        "operationId": "listMigrationCampaigns",
        "security": [
          {
            "workload": [],
            "actorDelegation": []
          }
        ],
        "parameters": [
          {
            "name": "site_id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "environment",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "resource_id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Scoped result; no native authority is conferred",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CampaignList"
                }
              }
            }
          },
          "default": {
            "description": "Held, denied or unavailable",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      },
      "post": {
        "operationId": "createMigrationCampaign",
        "security": [
          {
            "workload": [],
            "actorDelegation": []
          }
        ],
        "parameters": [
          {
            "name": "Idempotency-Key",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "202": {
            "description": "Scoped result; no native authority is conferred",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CampaignReceipt"
                }
              }
            }
          },
          "default": {
            "description": "Held, denied or unavailable",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateRequest"
              }
            }
          }
        }
      }
    },
    "/v1/tenants/{tenant}/migration-campaigns/{campaign}": {
      "parameters": [
        {
          "name": "tenant",
          "in": "path",
          "required": true,
          "schema": {
            "type": "string",
            "format": "uuid"
          }
        },
        {
          "name": "campaign",
          "in": "path",
          "required": true,
          "schema": {
            "type": "string",
            "format": "uuid"
          }
        }
      ],
      "get": {
        "operationId": "getMigrationCampaign",
        "security": [
          {
            "workload": [],
            "actorDelegation": []
          }
        ],
        "parameters": [],
        "responses": {
          "200": {
            "description": "Scoped result; no native authority is conferred",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Campaign"
                }
              }
            }
          },
          "default": {
            "description": "Held, denied or unavailable",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/v1/tenants/{tenant}/migration-campaigns/{campaign}/commands": {
      "parameters": [
        {
          "name": "tenant",
          "in": "path",
          "required": true,
          "schema": {
            "type": "string",
            "format": "uuid"
          }
        },
        {
          "name": "campaign",
          "in": "path",
          "required": true,
          "schema": {
            "type": "string",
            "format": "uuid"
          }
        }
      ],
      "post": {
        "operationId": "commandMigrationCampaign",
        "security": [
          {
            "workload": [],
            "actorDelegation": []
          }
        ],
        "parameters": [
          {
            "name": "Idempotency-Key",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "202": {
            "description": "Scoped result; no native authority is conferred",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CampaignReceipt"
                }
              }
            }
          },
          "default": {
            "description": "Held, denied or unavailable",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CommandRequest"
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "workload": {
        "type": "http",
        "scheme": "bearer"
      },
      "actorDelegation": {
        "type": "apiKey",
        "in": "header",
        "name": "X-Actor-Delegation"
      }
    },
    "schemas": {
...
      "MemberRequest": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "plan_id": {
            "type": "string",
            "format": "uuid"
          },
          "plan_revision": {
            "type": "integer",
            "minimum": 1,
            "maximum": 1152921504606846976
          },
          "plan_digest": {
            "type": "string",
            "pattern": "^[a-f0-9]{64}$"
          },
          "approval_id": {
            "type": "string",
            "format": "uuid"
          },
          "depends_on": {
            "type": "array",
            "items": {
              "type": "string",
              "format": "uuid"
            },
            "minItems": 0,
            "maxItems": 1000
          },
          "outage_group": {
            "oneOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ]
          },
          "priority": {
            "type": "integer",
            "minimum": 0,
            "maximum": 100
          },
          "not_before": {
            "type": "integer",
            "minimum": 0,
            "maximum": 1152921504606846976
          },
          "deadline": {
            "type": "integer",
            "minimum": 1,
            "maximum": 1152921504606846976
          }
        },
        "required": [
          "id",
          "plan_id",
          "plan_revision",
          "plan_digest",
          "approval_id",
          "depends_on",
          "outage_group",
          "priority",
          "not_before",
          "deadline"
        ]
      },
...
      "CreateRequest": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "scope": {
            "type": "object",
            "additionalProperties": false,
            "properties": {
              "site_id": {
                "type": "string",
                "format": "uuid"
              },
              "environment": {
                "type": "string",
                "format": "uuid"
              },
              "resource_id": {
                "type": "string",
                "format": "uuid"
              }
            },
            "required": [
              "site_id",
              "environment",
              "resource_id"
            ]
          },
          "settings": {
            "type": "object",
            "additionalProperties": false,
            "properties": {
              "name": {
                "type": "string",
                "minLength": 1,
                "maxLength": 120
              },
              "timezone": {
                "type": "string",
                "minLength": 1,
                "maxLength": 100
              },
              "windows": {
                "type": "array",
                "items": {
                  "type": "object",
                  "additionalProperties": false,
                  "properties": {
                    "start": {
                      "type": "integer",
                      "minimum": 0,
                      "maximum": 1152921504606846976
                    },
                    "end": {
                      "type": "integer",
                      "minimum": 1,
                      "maximum": 1152921504606846976
                    }
                  },
                  "required": [
                    "start",
                    "end"
                  ]
                },
                "minItems": 1,
                "maxItems": 128
              },
              "cutover_windows": {
                "type": "array",
                "items": {
                  "type": "object",
                  "additionalProperties": false,
                  "properties": {
                    "start": {
                      "type": "integer",
                      "minimum": 0,
                      "maximum": 1152921504606846976
                    },
                    "end": {
                      "type": "integer",
                      "minimum": 1,
                      "maximum": 1152921504606846976
                    }
                  },
                  "required": [
                    "start",
                    "end"
                  ]
                },
                "minItems": 1,
                "maxItems": 128
              },
              "blackouts": {
                "type": "array",
                "items": {
                  "type": "object",
                  "additionalProperties": false,
                  "properties": {
                    "start": {
                      "type": "integer",
                      "minimum": 0,
                      "maximum": 1152921504606846976
                    },
                    "end": {
                      "type": "integer",
                      "minimum": 1,
                      "maximum": 1152921504606846976
                    }
                  },
                  "required": [
                    "start",
                    "end"
                  ]
                },
                "minItems": 0,
                "maxItems": 128
              },
              "stagger_seconds": {
                "type": "integer",
                "minimum": 0,
                "maximum": 86400
              },
              "max_active": {
                "type": "integer",
                "minimum": 1,
                "maximum": 1000
              },
              "phase_limits": {
                "type": "object",
                "additionalProperties": false,
                "properties": {
                  "capture": {
                    "type": "integer",
                    "minimum": 1,
                    "maximum": 1000
                  },
                  "transfer": {
                    "type": "integer",
                    "minimum": 1,
                    "maximum": 1000
                  },
                  "conversion": {
                    "type": "integer",
                    "minimum": 1,
                    "maximum": 1000
                  },
                  "import": {
                    "type": "integer",
                    "minimum": 1,
                    "maximum": 1000
                  },
                  "validation": {
                    "type": "integer",
                    "minimum": 1,
                    "maximum": 1000
                  },
                  "cutover": {
                    "type": "integer",
                    "minimum": 1,
                    "maximum": 1000
                  }
                },
                "required": [
                  "capture",
                  "transfer",
                  "conversion",
                  "import",
                  "validation",
                  "cutover"
                ]
              },
              "failure_limit": {
                "type": "integer",
                "minimum": 1,
                "maximum": 1000
              }
            },
            "required": [
              "name",
              "timezone",
              "windows",
              "cutover_windows",
              "blackouts",
              "stagger_seconds",
              "max_active",
              "phase_limits",
              "failure_limit"
            ]
          },
          "members": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/MemberRequest"
            },
            "minItems": 1,
            "maxItems": 1000
          }
        },
        "required": [
          "scope",
          "settings",
          "members"
        ]
      },
      "CommandRequest": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "action": {
            "enum": [
              "schedule",
              "pause",
              "cancel"
            ]
          },
          "expected_revision": {
            "type": "integer",
            "minimum": 1,
            "maximum": 1152921504606846976
          }
        },
        "required": [
          "action",
          "expected_revision"
        ]
      },
      "CampaignReceipt": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "tenant_id": {
            "type": "string",
            "format": "uuid"
          },
          "scope": {
            "type": "object",
            "additionalProperties": false,
            "properties": {
              "site_id": {
                "type": "string",
                "format": "uuid"
              },
              "environment": {
                "type": "string",
                "format": "uuid"
              },
              "resource_id": {
                "type": "string",
                "format": "uuid"
              }
            },
            "required": [
              "site_id",
              "environment",
              "resource_id"
            ]
          },
          "revision": {
            "type": "integer",
            "minimum": 1,
            "maximum": 1152921504606846976
          },
          "state": {
            "enum": [
              "draft",
              "scheduled",
              "paused",
              "cancelled",
              "complete"
            ]
          }
        },
        "required": [
          "id",
          "tenant_id",
          "scope",
          "revision",
          "state"
        ]
      },
...
      "CampaignList": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "tenant_id": {
            "type": "string",
            "format": "uuid"
          },
          "scope": {
            "type": "object",
            "additionalProperties": false,
            "properties": {
              "site_id": {
                "type": "string",
                "format": "uuid"
              },
              "environment": {
                "type": "string",
                "format": "uuid"
              },
              "resource_id": {
                "type": "string",
                "format": "uuid"
              }
            },
            "required": [
              "site_id",
              "environment",
              "resource_id"
            ]
          },
          "items": {
            "type": "array",
            "items": {
              "type": "object",
              "additionalProperties": false,
              "properties": {
                "id": {
                  "type": "string",
                  "format": "uuid"
                },
                "state": {
                  "enum": [
                    "draft",
                    "scheduled",
                    "paused",
                    "cancelled",
                    "complete"
                  ]
                },
                "revision": {
                  "type": "integer",
                  "minimum": 1,
                  "maximum": 1152921504606846976
                },
                "name": {
                  "type": "string",
                  "minLength": 1,
                  "maxLength": 120
                },
                "created_at": {
                  "type": "integer",
                  "minimum": 0,
                  "maximum": 1152921504606846976
                }
              },
              "required": [
                "id",
                "state",
                "revision",
                "name",
                "created_at"
              ]
            },
            "minItems": 0,
            "maxItems": 100
          },
          "limit_reached": {
            "type": "boolean"
          }
        },
        "required": [
          "tenant_id",
          "scope",
          "items",
          "limit_reached"
        ]
      },
...
      "Error": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "error": {
            "type": "string",
            "pattern": "^[a-z_]{1,120}$"
          }
        },
        "required": [
          "error"
        ]
      }
    }
  }
}